requirements.txt
# Plotting/ Web Apps
matplotlib
plotly
dash
dash-bootstrap-components
ipykernel

# Data fetching and parsing
kaggle==1.5.12

# Machine learning essentials
scikit-learn
numpy
pandas
xgboost
nvidia_nccl_cu12==0


# others
nbformat
python-dotenv
